Relax aos event loop test latency thresholds
We are seeing on CI machines that hitting 50ms is not reasonable,
especially under high load.  --runs_per_test doesn't trigger this since
that doesn't load up the rest of the system heavily.

Efficiency isn't as big a concern as reliability.  Give ourselves much
more time (10x) for everything by just multiplying all the times by 10
so we are much more tolerant of a couple ms here and there.

Change-Id: I4d3c02760ec6bdafc76befc1a7ee9fcb4c271ef5
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>