Speed up downloading code.
Each SSH session takes time, so remove some.  Also, 3 sync's seems a bit
redundant...  Bring that down to 1.

The next thing to do is to create a folder locally with everything we
want in it (use symlinks and teach RSYNC to follow them) and get that
down to 2 SSH calls from the 4 we have today.

Change-Id: I62979a8a84953215cb9f9e72e8ea9fbb5a09c824
Signed-off-by: Austin Schuh <austin.linux@gmail.com>