Have downloader rsync everything at once for speed
This speeds up the copy *significantly*.  The extra SSH calls were
adding up to a lot of extra cost.  3 seconds for a NOP download!

Do this by building up a folder that looks like what the target should
look like, and then rsync that all in 1 go.

While we are here, rename .stripped to not have the extension.  It makes
the commands harder to remember.

Change-Id: I50e50c90421e049f48af453c8113d7c2204c7774
Signed-off-by: Austin Schuh <austin.linux@gmail.com>