Disable the Nagle algorithm for timestamps
We had left the Nagle algorithm for batching SCTP packets together to
reduce network bandwidth enabled on the client -> server direction.
This didn't affect data, but was affecting timestamp replies.  We use
those for timestamp estimation (and they are pretty small and
infrequent), so let's send them without delay instead.

Change-Id: I99cca1b8f9e59f74823d1021cab3803f5b0037f1
Signed-off-by: Austin Schuh <austin.linux@gmail.com>