Tune IRQs on the new roboRIO for lower latency
This pulls peak wakeup latency down from 800 uS or more to 150 uS.  Most
of this is changing ktimersoftd/* to be high priority.

While we are here, most of the IRQs, including MMC are running at RT
priority 15.  Anything we want to run needs to be higher than 15.  Move
the networking stack and message_bridge up above that so we get low
latency communications.

Change-Id: I14ede0fdeb2a008f482feae75fa04cabc3f5abeb
Signed-off-by: Austin Schuh <austin.linux@gmail.com>