Don't queue all future messages in memory in LogReader
When we were dropping messages at the end of the log file without data,
we were doing that by reading to the end of the file.  This actually
triggers all the rest of the nodes to queue up their data until that
point, which can cause a lot of memory to be used rather quickly.

Instead, just mark down that we shouldn't be sending and don't send.
This keeps things memory efficient.

Change-Id: I7417d4d5b720ab7986b5a55f9efba5612f962d58
Signed-off-by: Austin Schuh <austin.linux@gmail.com>