Make SkipTimingReport and DisableStatistics actually work
New event loops were not respecting the previous SkipTimingReports on
the factory.  Close that hole.

DisableStatistics was sending 1 statistics out at the beginning too.
Close that hole too.

I want to add a test that requires the following behavior to stay
working.  Let's test that this behavior continues to work as expected as
well in regards to the MessageBridgeServer and MessageBridgeClient and
Timestamp and RemoteMessage messages.

Change-Id: Ic1c51a14f9f76527bf1e8771f5247ba0aa4e94b9
Signed-off-by: Austin Schuh <austin.linux@gmail.com>