Handle start times before the start of a log file
We have seen log files where the start time is before any data after a
node reboots.  The side effect of this is that we try to call the
startup callback at the start time, but we only officially reboot at the
first message, so we can't actually start at the start time.  The world
then explodes.

Add a test reproducing it, and then fix it.

Change-Id: I7de986a6f7df0d050a0c8164827d27c799221fbf
Signed-off-by: Austin Schuh <austin.linux@gmail.com>