Error more gracefully in logfile sorting if 2 servers fight
If 2 computers have the same IP address, the client will flip flop
between them.  This could happen if 2 laptops are configured as the
logger node and are swapped back and forth, or various other patterns.

In this case, the timestamp problem is poorly formed and ends up failing
pretty horribly.  While we could try to figure out how to replay the
world during that time, in reality this isn't a problem we should be
solving.  We can catch this during logfile sorting and explode with a
much better error.

Connected to node A
                     Connected to node B
10sec - 20sec
                     20sec - 30sec
30sec - 40sec

In the above example, we are printing out the times on the logger node
which stays booted through this whole event.  If we track the times we
receive unreliable messages on the logger node, we can see that the
boots overlap.  A has time from 10-40, and B has time from 20-30.  So
there's overlap so we have this rebooting case.

Change-Id: I24746625c5fc4838a97ec7929cf1bd9ad6d7ffd6
Signed-off-by: Austin Schuh <austin.linux@gmail.com>