Track message_bridge client UUID and connection counts and time
It is hard to tell from the status message if a node has been
reconnecting a bunch or how long it has been connected.  Add both of
those numbers to both the client and server statistics.

While we are here, also add the server's boot UUID to the client.
Because of the way the protocol works, we only get this when the first
message is received, but that happens pretty quickly due to the
timestamps.

All of this should give us much better debugging around client/server
connections when things start to go south.

Change-Id: I2ed732afc81a045c7701fe47d1460a7df9c3d778
Signed-off-by: Austin Schuh <austin.linux@gmail.com>