Just memcpy data that might not be aligned
It'll be plenty fast for us. I suspect it's going to be just as fast
anywhere, but I'll argue about that with upstream later when I try
upstreaming this. The big advantage is it's definitely well-defined now,
vs before the attempt at using nullptr to make it work on hypothetical
C++ implementations that do weird things with pointers did not actually
do that, and triggered Clang warnings because it wasn't doing anything
useful.

Change-Id: I3c239b94b45449dc0a73c2c445778d72db8319df
Signed-off-by: Brian Silverman <bsilver16834@gmail.com>