Use remote sent times for calculating uptime in GetStatus
Do the same thing as when retrieving all statuses and use the monotonic
sent time of the status message as the "now" for calculating uptime.
This doesn't include the latency between the sending of the status
message and when aos_starter is run, but for typical use-cases that
shouldn't be an issue.

Change-Id: I73104b6a819d17817b4cbaf15378b91c852195bd
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>