Don't put reliable timestamps in the unreliable log file header field
James and I debugged a log file sorting bug which turned out to be a
logging bug.  The SystemParameters timestamp file (which we can all
agree is a reliable message) had the delivery time in
oldest_unreliable_remote_monotonic_timestamps, which is only for
unreliable timestamps.

This was because reliability was being tracked for the contents channel,
not the source message.  Instead, we need to do the book-keeping to
track if the source is realiable or not, and update the header
accordingly.  While we don't want to see new combined timestamp logs, we
still need to support them.  We don't mind if they are a bit less
efficient to log.

Change-Id: I292f5c0a64926904e13fb5d712acd523f3849cd4
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>