Adding a camera_id to our intrinsic calibration file
We'll need to manually label and track these, since the pi cameras
don't have an internal serial number (as best I can tell).

I think this is important, since we can swap cameras between pi's
and it's hard to be sure we've got the right calibration based on
pi hostame.

Including calibration files for cameras 22-01 to 22-07

Added a little detail to the CHECK, since it wasn't very clear

Change-Id: I6dd6f8e19355bb0fefce2fcb7fd79ba5c95fb374
Signed-off-by: Jim Ostrowski <yimmy13@gmail.com>