Add the catapult simulation and design code
This is mostly shamelessly stolen from rotory system.  It is different
enough because we have different control goals to not warrent being
lumped with the rotory system.

This lets us compute final speeds given the gear ratio and masses,
letting us optimize the gear ratio.

This gives us a model to start using when writing the optimization
problem for the fixed horizon MPC for shooting.

Change-Id: I03ac3448aa6ef717965fef7f41ef0a1197ea77a5
Signed-off-by: Austin Schuh <austin.linux@gmail.com>