Create a library for serving static files
It serves files that are in a given directory, and this could be
extended to html files.

In order to validate the functionality we also added a simple
`ScoutingServer` that you can `Start()` and `Stop()`.  The unit test
makes use of this. The `main` binary that ties all the components
together also makes use of this.

See the README for an overview of everything we added in this patch.

Change-Id: Id496b032d6aa70fd8502eeb347895ac52b5d1ddd
Signed-off-by: Het Satasiya <satasiyahet@gmail.com>
Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>