Deploy IMU driver and configure IRQs
The IRQs didn't match the existing IRQs.  Update them to be more
flexible and not depend on IRQ numbers since those appear to change.

wiringpi wasn't building for me, so I built the .deb and installed that
in the rootfs so I could make progress again.

While we are here, deploy to ~/bin instead of ~/robot_code/ so it is in
the path everywhere.  To make things actually easy to use, we need to
change how the config is found, but until then, this gets us closer to
easy to use.

Change-Id: Ib2f900f59bb6cc3de8f97e7687e723f6c8d003e7
Signed-off-by: Austin Schuh <austin.linux@gmail.com>