Look next to the binaries for config files.
This makes it so that we can run tools like aos_dump when they are put
in the path as long as the config ends up in the same folder as the
tool.  That's a much easier requirement.

At some point we probably need a better config deployment strategy, but
that time hasn't come and I don't think we have great requirements yet.

Change-Id: I4baf63d47998b0458d9e37a1619817337f08528f
Signed-off-by: Austin Schuh <austin.linux@gmail.com>