Add support for solving for turrets
Optionally, use turrent positions to apply a rotation around Z to model
the turret.  We then need to introduce a rotation from IMU frame to
turret frame (right before the bearing), and a translation as well.
This lets us represent the turret motion.

This is actually underconstrained.  We can pick arbitrary turret pivot
points along the turret rotation axis without changing the solution, and
can arbitrarily rotate 0 on the turret the same way.  Freeze Z in the
translation and apply a cost to the Z component of the rotation from IMU
to the turret frame to define these.

Change-Id: I3fed35a53e8e77e2c9feca9f5c6c3896b22c6277
Signed-off-by: Austin Schuh <austin.linux@gmail.com>