scouting: Add `defense_received_rating` to SubmitDataScouting
The web page that is in the works
(I467bd16caade9c987022600c2b63e9fad20da1a3) collects additional data
that SubmitDataScouting doesn't allow. The most obvious one to me is
the `defensePlayedOnScore` data. As far as I can tell, this is the
rating for the defense that other robots played on the robot of
interest.

This patch adds a `defense_received_rating` field to the
SubmitDataScouting message so that the web page can submit this
information.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I19d9067854854a9c9ba329f9b8c19a3152173518