Fill out AddToStats handler and test
This patch makes it so we can submit data scouting data to the
database. All tests are updated.

This patch also adds more error checking code to the database code to
make sure we get notified of some important edge cases.

The majority of the patch is written by Sabina. I amended the
`cli_test.py` file.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Signed-off-by: Sabina Leaver <100027607@mvla.net>
Change-Id: I9c70b8b02bbd3ff434fbbb857c2ed65f2684a50e