Add full error checking to //scouting/db
We had a few missing error checks in the code which made it hard to
debug the port to postgres. This patch adds the missing error checks
and cleans up the code a bit. I took this opportunity to rename
`error_` to `err` to match the rest of the code base.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Idc67c450079ba8aa97db1c1685a982d77ab8d44d