Undelay our status X_hat
There are 2 steps in the Kalman Filter.  Predict and Correct.  We
typically run Correct, compute our controller value, and then run
Predict.

Previously, we were grabbing the Predicted value on the cycle before it
was correct.  This means we were publishing our best guess at where the
state would be next cycle instead of the state we used.  This isn't
particularly helpful...

Change-Id: I03c1088d0b919e1fbf17ceabd1448d7f1894149a
Signed-off-by: Austin Schuh <austin.linux@gmail.com>