Simplify climbing screen on the scouting app
This patch gets rid of the secondary screen that popped up when
selecting "Attempted to Climb". It was confusing a few people. Now all
options are on the very first screen. Also, I decided to get rid of
all the `set*()` functions. It was simpler to use a two-way connection
for the radio buttons. I also got rid of the "proper" concept. I'm
still not sure what that was all about.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: Iaa9c2dda6bb5af2aa14dae50dbe20bcc96ef1bbe