Fix browser cache issues with the scouting app
Before this patch, folks had trouble getting the latest version of the
scouting app when visiting https://scouting.frc971.org/. The reason
for that is that they used the `Last-Modified` timestamp of the
individual files to determine whether they need to be fetched. Because
of `pkg_tar()`'s reproducibility, the timestamp is always the same.

This patch works around this by adding a `touch` step to the package's
`postinst`. This step updates the `Last-Modified` time of each
deployed file to the current time. In essence, every time the scouting
app is deployed, users should now see the latest version.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I6ba71c0a3ce791c0b281736397ff75c4375a1000