Simplify scouting bundle
The `babel()` rule dumps its outputs into a directory without telling
bazel about what files are in that directory. That means other rules
are forced to interact with the directory instead of the actual file
we're interested in.

This patch adds a `genrule()` to copy the one file we care about out
of that problematic directory. This lets other rules easily interact
with that one file.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I46233dd47f4bfdc77dfe562e95a0adab20aa20c5