Rework primary keys in the data scouting table
The ID is not actually used anywhere. The new scheme of using team
number, match number, round, and comp level as a unique key will let
us update data scouting data.

More importantly, this adds the missing round and comp level
information into the stats table. We need that to be able to collect
eliminations data.

Signed-off-by: Philipp Schrader <philipp.schrader@gmail.com>
Change-Id: I5614ada40a044a68bde168526ae65df47fd21db1